Fix compile issues after refactoring
