+++ src/main/scala/zio/zmx/metrics/StatsdInstrumentation.scala
@@ -1,6 +1,7 @@
 package zio.zmx.metrics
 
 import zio.ZIO
+import zio.zmx.MetricsDataModel._
 
 private[metrics] class StatsdInstrumentation extends ZMetrics.Service {
   override def counter(name: String): ZIO[Any, Nothing, Option[Metric.Counter]]        = ???
